pyproject.toml
[tool.poetry]
name = "tgchannelmanager"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^22.3"
pydantic-settings = "^2.10.1"
funcy = "^2.0"
structlog = "^25.4.0"
ecs-logging = "^2.2.0"
black = "^25.1.0"
ruff = "^0.12.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
